+++ routes/nqRoutes/list.js
@@ -8,6 +8,9 @@ const Image = require('../../models/nqModels/Image.js')
 const Video = require('../../models/nqModels/Video.js')
 const Article = require('../../models/nqModels/Article.js')
 const Note = require('../../models/nqModels/Note.js')
+const Doc = require('../../models/nqModels/Document.js')
+const Discourse = require('../../models/nqModels/Discourse.js')
+const Composition = require('../../models/nqModels/Composition.js')
 const Quote = require('../../models/nqModels/Quote.js')
 const Outline = require('../../models/nqModels/Outline.js')
 const Idea = require('../../models/nqModels/Idea.js')
@@ -20,6 +23,9 @@ var mediaList = {
     'videos': Video,
     'articles': Article,
     'notes': Note,
+    'document': Doc,
+    'discourse': Discourse,
+    'composition': Composition,
     'quotes': Quote,
     'outline': Outline,
     'idea': Idea,